# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AI Agent Productivity Platform - Open API",
        "contact": {},
        "version": "v0.0.1"
    },
    "basePath": "/openapi/v1",
    "paths": {
...
        "/rag/chat": {
            "post": {
                "description": "文本问答OpenAPI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openapi"
                ],
                "summary": "文本问答OpenAPI",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OpenAPIRagChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.OpenAPIRagChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.History": {
            "type": "object",
            "properties": {
                "needHistory": {
                    "type": "boolean"
                },
                "query": {
                    "type": "string"
                },
                "response": {
                    "type": "string"
                }
            }
        },
...
        "request.OpenAPIRagChatRequest": {
            "type": "object",
            "required": [
                "query",
                "uuid"
            ],
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.History"
                    }
                },
                "query": {
                    "type": "string"
                },
                "stream": {
                    "type": "boolean"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "response.OpenAIChatHistory": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string"
                },
                "response": {
                    "type": "string"
                }
            }
        },
        "response.OpenAIChatSearch": {
            "type": "object",
            "properties": {
                "kb_name": {
                    "type": "string"
                },
                "snippet": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "response.OpenAPIRagChatData": {
            "type": "object",
            "properties": {
                "output": {
                    "type": "string"
                },
                "searchList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.OpenAIChatSearch"
                    }
                }
            }
        },
        "response.OpenAPIRagChatResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/response.OpenAPIRagChatData"
                },
                "finish": {
                    "type": "integer"
                },
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.OpenAIChatHistory"
                    }
                },
                "message": {
                    "type": "string"
                },
                "msg_id": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}